+++ src/components/button/Button.js
@@ -64,15 +64,27 @@ export class ButtonComponent extends BaseComponent {
     return this.buttonElement;
   }
 
+  get emptyValue() {
+    return false;
+  }
+
   getValue() {
-    if (!this.component.input) {
-      return;
-    }
-    return this.clicked;
+    return this.dataValue;
+  }
+
+  get clicked() {
+    return this.dataValue;
   }
 
   get defaultValue() {
     return false;
+  }
+
+  set dataValue(value) {
+    if (!this.component.input) {
+      return;
+    }
+    super.dataValue = value;
   }
 
   get className() {
@@ -86,7 +98,7 @@ export class ButtonComponent extends BaseComponent {
       this.component.hidden = true;
     }
 
-    this.clicked = false;
+    this.dataValue = false;
     this.hasError = false;
     this.createElement();
     this.createInput(this.element);
@@ -130,7 +142,7 @@ export class ButtonComponent extends BaseComponent {
       }, true);
       this.on('change', (value) => {
         this.loading = false;
-        let isValid = this.root.isValid(value.data, true);
+        const isValid = this.root.isValid(value.data, true);
         this.disabled = this.options.readOnly || (this.component.disableOnInvalid && !isValid);
         if (isValid && this.hasError) {
           this.hasError = false;
@@ -162,7 +174,7 @@ export class ButtonComponent extends BaseComponent {
       }, true);
     }
     this.addEventListener(this.buttonElement, 'click', (event) => {
-      this.clicked = false;
+      this.dataValue = true;
       switch (this.component.action) {
         case 'submit':
           event.preventDefault();
@@ -241,11 +253,11 @@ export class ButtonComponent extends BaseComponent {
     }
 
     function getUrlParameter(name) {
-      name = name.replace(/[\[]/, '\\[').replace(/[\]]/, '\\]');
-      var regex = new RegExp('[\\?&]' + name + '=([^&#]*)');
+      name = name.replace(/[[]/, '\\[').replace(/[\]]/, '\\]');
+      var regex = new RegExp(`[\\?&]${name}=([^&#]*)`);
       var results = regex.exec(location.search);
       return results === null ? '' : decodeURIComponent(results[1].replace(/\+/g, ' '));
-    };
+    }
 
     // If this is an OpenID Provider initiated login, perform the click event immediately
     if (this.component.action === 'oauth' && this.component.oauth.authURI.indexOf(getUrlParameter('iss')) === 0) {