+++ src/components/currency/Currency.js
@@ -1,6 +1,7 @@
 import maskInput from 'vanilla-text-mask';
 import {createNumberMask} from 'text-mask-addons';
 import _ from 'lodash';
+import {getCurrencyAffixes} from '../../utils';
 
 import {NumberComponent} from '../number/Number';
 
@@ -25,39 +26,20 @@ export class CurrencyComponent extends NumberComponent {
   }
 
   constructor(component, options, data) {
+    // Currency should default to have a delimiter unless otherwise specified.
+    if (component && !component.hasOwnProperty('delimiter')) {
+      component.delimiter = true;
+    }
     super(component, options, data);
-
-    this.currency = this.component.currency || 'USD';
-    this.decimalLimit = this.component.decimalLimit || 2;
-
-    // Get the prefix and suffix from the localized string.
-    const regex = `(.*)?100(${this.decimalSeparator === '.'
-      ? '\\.'
-      : this.decimalSeparator}0{${this.decimalLimit}})?(.*)?`;
-    const parts = (100).toLocaleString(options.language, this.getFormatOptions()).match(new RegExp(regex));
-    this.prefix = parts[1] || '';
-    this.suffix = parts[3] || '';
-  }
-
-  getFormatOptions() {
-    return {
-      style: 'currency',
-      currency: this.currency,
-      useGrouping: true,
-      maximumFractionDigits: _.get(this.component, 'decimalLimit', this.decimalLimit)
-    };
-  }
-
-  formatNumber(value) {
-    try {
-      // Strip out the prefix and suffix before parsing. This occurs when numbers are from an old renderer.
-      value = value.replace(this.prefix, '').replace(this.suffix, '');
-    }
-    catch (e) {
-      // If value doesn't have a replace method, continue on as before.
-    }
-
-    return super.formatNumber(value);
+    this.decimalLimit = _.get(this.component, 'decimalLimit', 2);
+    const affixes = getCurrencyAffixes({
+      currency: this.component.currency,
+      decimalLimit: this.decimalLimit,
+      decimalSeparator: this.decimalSeparator,
+      lang: this.options.language,
+    });
+    this.prefix = affixes.prefix;
+    this.suffix = affixes.suffix;
   }
 
   parseNumber(value) {
@@ -68,17 +50,36 @@ export class CurrencyComponent extends NumberComponent {
   }
 
   setInputMask(input) {
-    this.inputMask = maskInput({
+    input.mask = maskInput({
       inputElement: input,
       mask: createNumberMask({
         prefix: this.prefix,
         suffix: this.suffix,
         thousandsSeparatorSymbol: _.get(this.component, 'thousandsSeparator', this.delimiter),
         decimalSymbol: _.get(this.component, 'decimalSymbol', this.decimalSeparator),
-        decimalLimit: _.get(this.component, 'decimalLimit', this.decimalLimit),
+        decimalLimit: this.decimalLimit,
         allowNegative: _.get(this.component, 'allowNegative', true),
         allowDecimal: _.get(this.component, 'allowDecimal', true)
       })
     });
+  }
+
+  clearInput(input) {
+    try {
+      input = input.replace(this.prefix, '').replace(this.suffix, '');
+    }
+    catch (err) {
+      // If value doesn't have a replace method, continue on as before.
+    }
+
+    return super.clearInput(input);
+  }
+
+  formatValue(value) {
+    if (this.component.requireDecimals && value && !value.includes(this.decimalSeparator)) {
+      return `${value}${this.decimalSeparator}${_.repeat('0', this.decimalLimit)}`;
+    }
+
+    return value;
   }
 }