+++ src/components/currency/Currency.spec.js
@@ -11,14 +11,14 @@ describe('Currency Component', () => {
   });
   it('Should format currency for USA locale', (done) => {
     Harness.testCreate(CurrencyComponent, comps.comp1, {language: 'en-US'}).then((component) => {
-      Harness.testSetInput(component, null, null, '');
-      Harness.testSetInput(component, undefined, null, '');
-      Harness.testSetInput(component, {}, null, '');
-      Harness.testSetInput(component, [], null, '');
-      Harness.testSetInput(component, [''], null, '');
+      Harness.testSetInput(component, null, 0, '$0');
+      Harness.testSetInput(component, undefined, 0, '$0');
+      Harness.testSetInput(component, {}, 0, '$0');
+      Harness.testSetInput(component, [], 0, '$0');
+      Harness.testSetInput(component, [''], 0, '$0');
       Harness.testSetInput(component, ['1'], 1, '$1.00');
       Harness.testSetInput(component, ['$1.00'], 1, '$1.00');
-      Harness.testSetInput(component, 0, 0, '0');
+      Harness.testSetInput(component, 0, 0, '$0.00');
       Harness.testSetInput(component, 1.00, 1, '$1.00');
       Harness.testSetInput(component, -1.00, -1, '-$1.00');
       Harness.testSetInput(component, 1, 1, '$1.00');
@@ -62,8 +62,8 @@ describe('Currency Component', () => {
   });
   it('Should format currency for British locale', (done) => {
     Harness.testCreate(CurrencyComponent, comps.comp1, {language: 'en-GB'}).then((component) => {
-      Harness.testSetInput(component, null, null, '');
-      Harness.testSetInput(component, 0, 0, '0');
+      Harness.testSetInput(component, null, 0, 'US$0');
+      Harness.testSetInput(component, 0, 0, 'US$0.00');
       Harness.testSetInput(component, 1.00, 1, 'US$1.00');
       Harness.testSetInput(component, -1.00, -1, '-US$1.00');
       Harness.testSetInput(component, 1,  1, 'US$1.00');
@@ -82,8 +82,8 @@ describe('Currency Component', () => {
   it('Should format currency for French locale', (done) => {
     Harness.testCreate(CurrencyComponent, comps.comp1, {language: 'fr'}).then((component) => {
       // The spaces in these tests are a weird unicode space so be careful duplicating the tests.
-      Harness.testSetInput(component, null, null, '');
-      Harness.testSetInput(component, 0, 0, '0');
+      Harness.testSetInput(component, null, 0, '0 $US');
+      Harness.testSetInput(component, 0, 0, '0,00 $US');
       Harness.testSetInput(component, 1.00, 1, '1,00 $US');
       Harness.testSetInput(component, -1.00, -1, '-1,00 $US');
       Harness.testSetInput(component, 1, 1, '1,00 $US');
@@ -101,8 +101,8 @@ describe('Currency Component', () => {
   });
   it('Should format currency for German locale', (done) => {
     Harness.testCreate(CurrencyComponent, comps.comp1, {language: 'de'}).then((component) => {
-      Harness.testSetInput(component, null, null, '');
-      Harness.testSetInput(component, 0, 0, '0');
+      Harness.testSetInput(component, null, 0, '0 $');
+      Harness.testSetInput(component, 0, 0, '0,00 $');
       Harness.testSetInput(component, 1.00, 1.00, '1,00 $');
       Harness.testSetInput(component, -1.00, -1.00, '-1,00 $');
       Harness.testSetInput(component, 1, 1, '1,00 $');