+++ src/components/datetime/DateTime.js
@@ -76,6 +76,10 @@ export class DateTimeComponent extends BaseComponent {
     return info;
   }
 
+  get emptyValue() {
+    return 0;
+  }
+
   /**
    * Get the default date for the calendar.
    * @return {*}
@@ -180,7 +184,7 @@ export class DateTimeComponent extends BaseComponent {
     return (new Date(timestamp * 1000));
   }
 
-  get validateValue() {
+  get validationValue() {
     const values = [];
     for (const i in this.inputs) {
       if (this.inputs.hasOwnProperty(i)) {