+++ src/formio.form.js
@@ -5,8 +5,6 @@ import Formio from './formio';
 import Promise from 'native-promise-only';
 import {FormioComponents} from './components/Components';
 
-i18next.initialized = false;
-
 // Initialize the available forms.
 Formio.forms = {};
 
@@ -211,10 +209,9 @@ export default class FormioForm extends FormioComponents {
     this.shortcuts = [];
 
     // Set language after everything is established.
-    if (options && options.language) {
-      i18n.lng = options.language;
-      this.language = options.language;
-    }
+    this.localize().then(() => {
+      this.language = this.options.language;
+    });
   }
 
   /**
@@ -735,18 +732,16 @@ export default class FormioForm extends FormioComponents {
   render() {
     return this.onElement.then(() => {
       this.clear();
-      //this.showElement(false);
-      return this.localize().then(() => {
-        this.build();
-        this.isBuilt = true;
-        this.onResize();
-        this.on('resetForm', () => this.reset(), true);
-        this.on('refreshData', () => this.updateValue());
-        setTimeout(() => {
-          this.onChange();
-          this.emit('render');
-        }, 1);
-      });
+      this.showElement(false);
+      this.build();
+      this.isBuilt = true;
+      this.onResize();
+      this.on('resetForm', () => this.reset(), true);
+      this.on('refreshData', () => this.updateValue());
+      setTimeout(() => {
+        this.onChange();
+        this.emit('render');
+      }, 1);
     });
   }