+++ src/formio.polyfill.js
@@ -2,7 +2,7 @@
  * DO NOT DELETE THIS! THIS WILL BREAK THE PDF RENDERING IF YOU DO.
  */
 
-"use strict";
+'use strict';
 
 /**
  * Taken from https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Function/bind
@@ -24,7 +24,7 @@ if (!Function.prototype.bind) {
         return fToBind.apply(this instanceof fNOP
           ? this
           : oThis,
-          aArgs.concat(Array.prototype.slice.call(arguments)));
+        aArgs.concat(Array.prototype.slice.call(arguments)));
       };
 
     if (this.prototype) {